# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Web-studio API docs",
    "description": "API documentation for web-studio-backend. [Backend github repository](https://github.com/IvanVA13/web-studio-backend)"
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "servers": [
    { "url": "https://web-studioiv.herokuapp.com/api" },
    { "url": "http://localhost:3000/api" }
  ],
...
  "paths": {
...
    "/users/verify/{verifyEmailToken}": {
      "get": {
        "tags": ["User"],
        "summary": "Verify user email",
        "parameters": [
          {
            "name": "verifyEmailToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token issued to verify user email"
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "404": {
            "description": "Your email has already been verified or User not found",
            "content": {}
          }
        }
      }
    },
...
    "/users/reset-password/{resetPasswordToken}": {
      "post": {
        "tags": ["User"],
        "summary": "User reset password",
        "parameters": [
          {
            "name": "resetPasswordToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reset password token issued to the user"
          }
        ],
        "requestBody": {
          "description": "Reset's object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "400": {
            "description": "missing required fields",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": {
            "type": "string",
            "description": "User's password",
            "example": "qwerty456"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}